requirements.txt
boto3==1.19.12
botocore==1.22.12
colorize==1.1.0
cycler==0.11.0
jmespath==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.4
opencv-python==4.5.4.58
Pillow==8.4.0
pyk4a==1.3.0
pyparsing==3.0.4
python-dateutil==2.8.2
python-decouple==3.5
python-version==0.0.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
